pyazuread/requirements.txt
requests==2.24.0
cryptography==3.4.8
PyJWT==1.7.1